# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /start:
    post:
      operationId: startProcessInstance
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessInstanceInputMessage'
      - in: query
        name: gid
        description: The UUID of the process instance group owned by the party initiating the process
        required: false
        type: string
      - in: query
        name: preceding_gid
        description: The UUID of the preceding process instance group
        required: false
        type: string
      - in: query
        name: collaborationGID
        description: The id of the collaboration group which the process instance group belongs to
        required: false
        type: string
      responses:
        # Response code
        200:
          description: Business Process Instance Started
          schema:
            $ref: '#/definitions/ProcessInstance'
...
  /continue:
    post:
      operationId: continueProcessInstance
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessInstanceInputMessage'
      - in: query
        name: gid
        description: The UUID of the process instance group owned by the party continuing the process
        required: false
        type: string
      - in: query
        name: collaborationGID
        description: The id of the collaboration group which the process instance group belongs to
        required: false
        type: string
      - name: Authorization
        in: header
        type: string
        required: true
      responses:
        # Response code
        200:
          description: Business Process Instance Message sent successfully
          schema:
            $ref: '#/definitions/ProcessInstance'

#  /document:
#    post:
#      operationId: addDocumentMetadata
#      consumes:
#      - application/json
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: body
#        name: body
#        required: true
#        schema:
#          $ref: '#/definitions/ProcessDocumentMetadata'
#      responses:
#        # Response code
#        200:
#          description: Document added
#          schema:
#            $ref: '#/definitions/ApiResponse'
#    put:
#      operationId: updateDocumentMetadata
#      consumes:
#      - application/json
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: body
#        name: body
#        schema:
#          $ref: '#/definitions/ProcessDocumentMetadata'
#      responses:
#        # Response code
#        200:
#          description: Document updated
#          schema:
#            $ref: '#/definitions/ApiResponse'
#
#  /document/{documentID}:
#    delete:
#      operationId: deleteDocument
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: path
#        name: documentID
#        type: string
#        required: true
#      responses:
#        # Response code
#        200:
#          description: Document deleted
#          schema:
#            $ref: '#/definitions/ApiResponse'
#
#  /document/{partnerID}/{type}/{source}/{status}:
#    get:
#      operationId: getDocuments
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: path
#        name: partnerID
#        type: string
#        description: The id of the partner.
#        required: true
#      - in: path
#        name: type
#        type: string
#        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
#        required: true
#      - in: path
#        name: status
#        type: string
#        description: The status of the document. (APPROVED, WAITINGRESPONSE, DENIED or PASTDUE)
#        required: true
#      - in: path
#        name: source
#        type: string
#        description: The source of the document (whether it is sent or received by the partner - SENT, RECEIVED).
#        required: true
#      responses:
#        # Response code
#        200:
#          description: Document obtained
#          schema:
#            type: array
#            items:
#              $ref: '#/definitions/ProcessDocumentMetadata'
#
#  /document/{partnerID}/{type}/{source}:
#    get:
#      operationId: getDocuments
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: path
#        name: partnerID
#        type: string
#        description: The id of the partner.
#        required: true
#      - in: path
#        name: type
#        type: string
#        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
#        required: true
#      - in: path
#        name: source
#        type: string
#        description: The source of the document (whether it is sent or received by the partner - SENT, RECEIVED).
#        required: true
#      responses:
#        # Response code
#        200:
#          description: Document obtained
#          schema:
#            type: array
#            items:
#              $ref: '#/definitions/ProcessDocumentMetadata'
#
#  /document/{partnerID}/{type}:
#    get:
#      operationId: getDocuments
#      produces:
#      - application/json
#      parameters:
#      - name: Authorization
#        in: header
#        type: string
#        required: true
#      - in: path
#        name: partnerID
#        type: string
#        description: The id of the partner.
#        required: true
#      - in: path
#        name: type
#        type: string
#        description: The type of the document such as PPAPRESPONSE,ORDERRESPONSESIMPLE,REQUESTFORQUOTATION etc.
#        required: true
#      responses:
#        # Response code
#        200:
#          description: Document obtained
#          schema:
#            type: array
#            items:
#              $ref: '#/definitions/ProcessDocumentMetadata'
...
definitions:
...
  ProcessVariables:
    title: Business Process Variables Object
    type: object
    properties:
      processID:
        type: string
      initiatorID:
        type: string
      responderID:
        type: string
      contentUUID:
        type: string
        description: The UUID of the document in the content variable
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
      content:
        type: string
      creatorUserID:
        type: string
  ProcessInstanceInputMessage:
    title: Business Process Input Message Object
    type: object
    properties:
      variables:
        $ref: '#/definitions/ProcessVariables'
      processInstanceID:
        type: string
  ProcessInstance:
    title: Business Process Instance
    type: object
    properties:
      processInstanceID:
        type: string
      creationDate:
        type: string
      processID:
        type: string
      status:
        type: string
        enum:
        - STARTED
        - CANCELLED
        - COMPLETED
...
  ProcessDocumentMetadata:
    title: Business Process Document Metadata Object
    type: object
    properties:
      documentID:
        type: string
      submissionDate:
        type: string
      type:
        type: string
        enum:
        - CATALOGUE
        - REQUESTFORQUOTATION
        - QUOTATION
        - ORDER
        - ORDERRESPONSESIMPLE
        - REMITTANCEADVICE
        - INVOICE
        - APPLICATIONRESPONSE
        - TRANSPORTATIONSTATUS
        - DESPATCHADVICE
        - RECEIPTADVICE
        - PPAPREQUEST
        - PPAPRESPONSE
        - TRANSPORTEXECUTIONPLANREQUEST
        - TRANSPORTEXECUTIONPLAN
        - ITEMINFORMATIONREQUEST
        - ITEMINFORMATIONRESPONSE
        - OTHER
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
      status:
        type: string
        enum:
        - APPROVED
        - WAITINGRESPONSE
        - DENIED
        - PASTDUE
      processInstanceID:
        type: string
      initiatorID:
        type: string
      responderID:
        type: string
      creatorUserID:
        type: string
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
      type:
        type: string
      message:
        type: string